summary CMIP5 files information

get_ensemble: Extract CMIP5 ensemle name from file names

extract_model: Extract CMIP5 model name from file names

CMIP5Files_info: Extract CMIP5 information from file names

CMIP5Files_filter: Filter corresponding duration or period files. Dates has been adjust from 12-01 to 01-01 for start_adj. Duplicated dates are also removed.

CMIP5Files_summary: Get the start and end information of every model

get_ensemble(files)

extract_model(files)

CMIP5Files_info(files)

CMIP5Files_summary(d)

CMIP5Files_filter(
  files,
  duration = 200,
  period = NULL,
  check_dupli = TRUE,
  verbose = 1
)

Arguments

files

CMIP5 nc files, full name path.

d

An object returned by CMIP5Files_filter or CMIP5Files_info or a data.frame at least with the columns of 'model', 'year_start_adj', 'year_end', 'file'.

duration

length of year.

period

starting and ending year, e.g. c(1850, 2005). If provided, duration will be overwritten. If duration and period are null at the same time, no filter applied.

check_dupli

Boolean. If true, duplicated date will be removed by check_dfile().

verbose

echo nc file missing info, if verbose >= 1

echo duplicated info, if verbose >= 2

Value

d_files A data.table with colnames of 'Id', 'model', 'kind', 'start', 'end', 'year_start', 'year_start_adj', 'year_end', 'file'

Note

Only one scenario per time.

See also

Examples

data(files_short)
ensemble <- get_ensemble(files_short)
model <- extract_model(files_short)
CMIP5Files_info(files_short)
#>         Id  scenario     model ensemble freq      start        end  start_adj
#>    1:    1 piControl ACCESS1-0   r1i1p1  day  300-01-01  324-12-31  300-01-01
#>    2:    2 piControl ACCESS1-0   r1i1p1  day  325-01-01  349-12-31  325-01-01
#>    3:    3 piControl ACCESS1-0   r1i1p1  day  350-01-01  374-12-31  350-01-01
#>    4:    4 piControl ACCESS1-0   r1i1p1  day  375-01-01  399-12-31  375-01-01
#>    5:    5 piControl ACCESS1-0   r1i1p1  day  400-01-01  424-12-31  400-01-01
#>   ---                                                                        
#> 2448: 2448 piControl NorESM1-M   r1i1p1  day 1000-01-01 1049-12-31 1000-01-01
#> 2449: 2449 piControl NorESM1-M   r1i1p1  day 1050-01-01 1099-12-31 1050-01-01
#> 2450: 2450 piControl NorESM1-M   r1i1p1  day 1100-01-01 1149-12-31 1100-01-01
#> 2451: 2451 piControl NorESM1-M   r1i1p1  day 1150-01-01 1150-12-31 1150-01-01
#> 2452: 2452 piControl NorESM1-M   r1i1p1  day 1151-01-01 1200-12-31 1151-01-01
#>          end_adj year_start year_end len
#>    1:  324-12-31        300      324  25
#>    2:  349-12-31        325      349  25
#>    3:  374-12-31        350      374  25
#>    4:  399-12-31        375      399  25
#>    5:  424-12-31        400      424  25
#>   ---                                   
#> 2448: 1049-12-31       1000     1049  50
#> 2449: 1099-12-31       1050     1099  50
#> 2450: 1149-12-31       1100     1149  50
#> 2451: 1150-12-31       1150     1150   1
#> 2452: 1200-12-31       1151     1200  50
#>                                                             file
#>    1: tasmax_day_ACCESS1-0_piControl_r1i1p1_03000101-03241231.nc
#>    2: tasmax_day_ACCESS1-0_piControl_r1i1p1_03250101-03491231.nc
#>    3: tasmax_day_ACCESS1-0_piControl_r1i1p1_03500101-03741231.nc
#>    4: tasmax_day_ACCESS1-0_piControl_r1i1p1_03750101-03991231.nc
#>    5: tasmax_day_ACCESS1-0_piControl_r1i1p1_04000101-04241231.nc
#>   ---                                                           
#> 2448: tasmax_day_NorESM1-M_piControl_r1i1p1_10000101-10491231.nc
#> 2449: tasmax_day_NorESM1-M_piControl_r1i1p1_10500101-10991231.nc
#> 2450: tasmax_day_NorESM1-M_piControl_r1i1p1_11000101-11491231.nc
#> 2451: tasmax_day_NorESM1-M_piControl_r1i1p1_11500101-11501231.nc
#> 2452: tasmax_day_NorESM1-M_piControl_r1i1p1_11510101-12001231.nc

# Get the summary infomation of CMIP5Files_filter or CMIP5Files_info
CMIP5Files_filter(files_short) %>% CMIP5Files_summary()
#> =========================== piControl  ===========================
#> 
#>              model ensemble  scenario      start        end  start_adj
#>  1:      ACCESS1-0   r1i1p1 piControl  600-01-01  799-12-31  600-01-01
#>  2:      ACCESS1-3   r1i1p1 piControl  550-01-01  749-12-31  550-01-01
#>  3:     bcc-csm1-1   r1i1p1 piControl  168-01-01  500-12-31  301-01-01
#>  4:   bcc-csm1-1-m   r1i1p1 piControl  201-01-01  400-12-31  201-01-01
#>  5:        BNU-ESM   r1i1p1 piControl 1450-01-01 1972-12-31 1450-01-01
#>  6:        CanESM2   r1i1p1 piControl 2911-01-01 3110-12-31 2911-01-01
#>  7:      CESM1-BGC   r1i1p1 piControl  400-01-01  600-12-31  401-01-01
#>  8:      CMCC-CESM   r1i1p1 piControl 4396-01-01 4600-12-31 4401-01-01
#>  9:        CMCC-CM   r1i1p1 piControl 1680-01-01 1879-12-31 1680-01-01
#> 10:       CMCC-CMS   r1i1p1 piControl 3984-01-01 4183-12-31 3984-01-01
#> 11:       CNRM-CM5   r1i1p1 piControl 2500-01-01 2699-12-31 2500-01-01
#> 12:  CSIRO-Mk3-6-0   r1i1p1 piControl  301-01-01  500-12-31  301-01-01
#> 13:       EC-EARTH   r1i1p1 piControl 2350-01-01 2551-12-31 2352-01-01
#> 14:      FGOALS-g2   r1i1p1 piControl  560-01-01  641-01-15  560-01-01
#> 15:      FGOALS-s2   r1i1p1 piControl 2151-01-01 2350-12-31 2151-01-01
#> 16:       GFDL-CM3   r1i1p1 piControl  601-01-01  800-12-31  601-01-01
#> 17:     GFDL-ESM2G   r1i1p1 piControl  301-01-01  500-12-31  301-01-01
#> 18:     GFDL-ESM2M   r1i1p1 piControl  301-01-01  500-12-31  301-01-01
#> 19:     HadGEM2-CC   r1i1p1 piControl 1899-12-01 2099-12-30 1900-01-01
#> 20:     HadGEM2-ES   r1i1p1 piControl 2228-12-01 2435-11-30 2236-01-01
#> 21:         inmcm4   r1i1p1 piControl 2150-01-01 2349-12-31 2150-01-01
#> 22:   IPSL-CM5A-LR   r1i1p1 piControl 2600-01-01 2799-12-31 2600-01-01
#> 23:   IPSL-CM5A-MR   r1i1p1 piControl 1900-01-01 2099-12-31 1900-01-01
#> 24:   IPSL-CM5B-LR   r1i1p1 piControl 1830-01-01 2129-12-31 1930-01-01
#> 25:      MIROC-ESM   r1i1p1 piControl 2120-01-01 2429-12-31 2230-01-01
#> 26: MIROC-ESM-CHEM   r1i1p1 piControl 1846-01-01 2100-12-31 1901-01-01
#> 27:        MIROC4h   r1i1p1 piControl   51-01-01  150-12-31   51-01-01
#> 28:         MIROC5   r1i1p1 piControl 2670-01-01 2869-12-31 2670-01-01
#> 29:     MPI-ESM-LR   r1i1p1 piControl 2650-01-01 2849-12-31 2650-01-01
#> 30:     MPI-ESM-MR   r1i1p1 piControl 2650-01-01 2849-12-31 2650-01-01
#> 31:      MPI-ESM-P   r1i1p1 piControl 2800-01-01 3005-12-31 2806-01-01
#> 32:      MRI-CGCM3   r1i1p1 piControl 2151-01-01 2350-12-31 2151-01-01
#> 33:      NorESM1-M   r1i1p1 piControl 1000-01-01 1200-12-31 1001-01-01
#>              model ensemble  scenario      start        end  start_adj
#>        end_adj year_start year_end   n len len_adj
#>  1:  799-12-31        600      799   8 200     200
#>  2:  749-12-31        550      749   8 200     200
#>  3:  500-12-31        168      500   2 333     200
#>  4:  400-12-31        201      400   8 200     200
#>  5: 1649-12-31       1450     1972   1 523     200
#>  6: 3110-12-31       2911     3110   2 200     200
#>  7:  600-12-31        400      600   8 201     200
#>  8: 4600-12-31       4396     4600  35 205     200
#>  9: 1879-12-31       1680     1879 200 200     200
#> 10: 4183-12-31       3984     4183  20 200     200
#> 11: 2699-12-31       2500     2699  40 200     200
#> 12:  500-12-31        301      500  10 200     200
#> 13: 2551-12-31       2350     2551  21 202     200
#> 14:  641-01-15        560      641  82  82      82
#> 15: 2350-12-31       2151     2350 200 200     200
#> 16:  800-12-31        601      800  40 200     200
#> 17:  500-12-31        301      500  40 200     200
#> 18:  500-12-31        301      500  40 200     200
#> 19: 2099-12-30       1899     2099  41 201     200
#> 20: 2435-11-30       2228     2435  21 208     200
#> 21: 2349-12-31       2150     2349  20 200     200
#> 22: 2799-12-31       2600     2799   1 200     200
#> 23: 2099-12-31       1900     2099  10 200     200
#> 24: 2129-12-31       1830     2129   2 300     200
#> 25: 2429-12-31       2120     2429   3 310     200
#> 26: 2100-12-31       1846     2100   2 255     200
#> 27:  150-12-31         51      150 100 100     100
#> 28: 2869-12-31       2670     2869  20 200     200
#> 29: 2849-12-31       2650     2849  20 200     200
#> 30: 2849-12-31       2650     2849  20 200     200
#> 31: 3005-12-31       2800     3005  19 206     200
#> 32: 2350-12-31       2151     2350  20 200     200
#> 33: 1200-12-31       1000     1200   5 201     200
#>        end_adj year_start year_end   n len len_adj

# filter CMIP5 files
CMIP5Files_filter(files_short, duration = 200)
#> =========================== piControl  ===========================
#> 
#>         Id  scenario     model ensemble freq      start        end  start_adj
#>    1:   13 piControl ACCESS1-0   r1i1p1  day  600-01-01  624-12-31  600-01-01
#>    2:   14 piControl ACCESS1-0   r1i1p1  day  625-01-01  649-12-31  625-01-01
#>    3:   15 piControl ACCESS1-0   r1i1p1  day  650-01-01  674-12-31  650-01-01
#>    4:   16 piControl ACCESS1-0   r1i1p1  day  675-01-01  699-12-31  675-01-01
#>    5:   17 piControl ACCESS1-0   r1i1p1  day  700-01-01  724-12-31  700-01-01
#>   ---                                                                        
#> 1065: 1856 piControl    inmcm4   r1i1p1  day 2300-01-01 2309-12-31 2300-01-01
#> 1066: 1857 piControl    inmcm4   r1i1p1  day 2310-01-01 2319-12-31 2310-01-01
#> 1067: 1858 piControl    inmcm4   r1i1p1  day 2320-01-01 2329-12-31 2320-01-01
#> 1068: 1859 piControl    inmcm4   r1i1p1  day 2330-01-01 2339-12-31 2330-01-01
#> 1069: 1860 piControl    inmcm4   r1i1p1  day 2340-01-01 2349-12-31 2340-01-01
#>          end_adj year_start year_end len
#>    1:  624-12-31        600      624  25
#>    2:  649-12-31        625      649  25
#>    3:  674-12-31        650      674  25
#>    4:  699-12-31        675      699  25
#>    5:  724-12-31        700      724  25
#>   ---                                   
#> 1065: 2309-12-31       2300     2309  10
#> 1066: 2319-12-31       2310     2319  10
#> 1067: 2329-12-31       2320     2329  10
#> 1068: 2339-12-31       2330     2339  10
#> 1069: 2349-12-31       2340     2349  10
#>                                                             file
#>    1: tasmax_day_ACCESS1-0_piControl_r1i1p1_06000101-06241231.nc
#>    2: tasmax_day_ACCESS1-0_piControl_r1i1p1_06250101-06491231.nc
#>    3: tasmax_day_ACCESS1-0_piControl_r1i1p1_06500101-06741231.nc
#>    4: tasmax_day_ACCESS1-0_piControl_r1i1p1_06750101-06991231.nc
#>    5: tasmax_day_ACCESS1-0_piControl_r1i1p1_07000101-07241231.nc
#>   ---                                                           
#> 1065:    tasmax_day_inmcm4_piControl_r1i1p1_23000101-23091231.nc
#> 1066:    tasmax_day_inmcm4_piControl_r1i1p1_23100101-23191231.nc
#> 1067:    tasmax_day_inmcm4_piControl_r1i1p1_23200101-23291231.nc
#> 1068:    tasmax_day_inmcm4_piControl_r1i1p1_23300101-23391231.nc
#> 1069:    tasmax_day_inmcm4_piControl_r1i1p1_23400101-23491231.nc