Get data nodes which store CMIP6 output

Get data nodes which store CMIP6 output

get_data_node(speed_test = TRUE, timeout = 3)

Arguments

speed_test

If TRUE, use pingr::ping() to perform connection speed test on each data node. A ping column is appended in returned data.table which stores each data node response in milliseconds. This feature needs pingr package already installed. Default: FALSE.

timeout

Timeout for a ping response in seconds. Default: 3.

Value

A data.table::data.table() of 2 or 3 (when speed_test is TRUE) columns:

Column	Type	Description

data_node	character	Web address of data node

status	character	Status of data node. "UP" means OK and "DOWN" means currently not available

ping	double	Data node response in milliseconds during speed test

Examples

# \donttest{
get_data_node()
#> ping: aims3.llnl.gov            NaN
#> ping: cmip.dess.tsinghua.edu.cn NaN
#> ping: cmip.fio.org.cn           NaN
#> ping: cordexesg.dmi.dk          NaN
#> ping: crd-esgf-drc.ec.gc.ca     NaN
#> ping: data.meteo.unican.es      NaN
#> ping: dataserver.nccs.nasa.gov  NaN
#> ping: dpesgf03.nccs.nasa.gov    NaN
#> ping: esg-cccr.tropmet.res.in   NaN
#> ping: esg-dn1.nsc.liu.se        NaN
#> ping: esg-dn1.ru.ac.th          NaN
#> ping: esg-dn3.nsc.liu.se        NaN
#> ping: esg.camscma.cn            NaN
#> ping: esg.lasg.ac.cn            NaN
#> ping: esg1.umr-cnrm.fr          NaN
#> ping: esg2.umr-cnrm.fr          NaN
#> ping: esgdata.gfdl.noaa.gov     NaN
#> ping: esgf-cnr.hpc.cineca.it    NaN
#> ping: esgf-data.ucar.edu        NaN
#> ping: esgf-data01.diasjp.net    NaN
#> ping: esgf-data02.diasjp.net    NaN
#> ping: esgf-data03.diasjp.net    NaN
#> ping: esgf-data04.diasjp.net    NaN
#> ping: esgf-data1.diasjp.net     NaN
#> ping: esgf-data1.llnl.gov       NaN
#> ping: esgf-data2.llnl.gov       NaN
#> ping: esgf-ictp.hpc.cineca.it   NaN
#> ping: esgf-nimscmip6.apcc21.org NaN
#> ping: esgf-node.cmcc.it         NaN
#> ping: esgf-node.ornl.gov        NaN
#> ping: esgf-node2.cmcc.it        NaN
#> ping: esgf.apcc21.org           NaN
#> ping: esgf.bsc.es               NaN
#> ping: esgf.ceda.ac.uk           NaN
#> ping: esgf.dwd.de               NaN
#> ping: esgf.ichec.ie             NaN
#> ping: esgf.nccs.nasa.gov        NaN
#> ping: esgf.nci.org.au           NaN
#> ping: esgf.rcec.sinica.edu.tw   NaN
#> ping: esgf1.dkrz.de             NaN
#> ping: esgf2.dkrz.de             NaN
#> ping: esgf3.dkrz.de             NaN
#> ping: noresg.nird.sigma2.no     NaN
#> ping: polaris.pknu.ac.kr        NaN
#> ping: vesg.ipsl.upmc.fr         NaN
#> ping: cmip.bcc.cma.cn           NaN
#> ping: dist.nmlab.snu.ac.kr      NaN
#> ping: dmiesgf0.dmi.dk           NaN
#> ping: eridanus.eoc.dlr.de       NaN
#> ping: esg-dn2.nsc.liu.se        NaN
#> ping: esg.pik-potsdam.de        NaN
#> ping: esgf-data.csc.fi          NaN
#> ping: esgf.anl.gov              NaN
#>                     data_node status speed_ms
#>  1:            aims3.llnl.gov     UP     9999
#>  2: cmip.dess.tsinghua.edu.cn     UP     9999
#>  3:           cmip.fio.org.cn     UP     9999
#>  4:          cordexesg.dmi.dk     UP     9999
#>  5:     crd-esgf-drc.ec.gc.ca     UP     9999
#>  6:      data.meteo.unican.es     UP     9999
#>  7:  dataserver.nccs.nasa.gov     UP     9999
#>  8:    dpesgf03.nccs.nasa.gov     UP     9999
#>  9:   esg-cccr.tropmet.res.in     UP     9999
#> 10:        esg-dn1.nsc.liu.se     UP     9999
#> 11:          esg-dn1.ru.ac.th     UP     9999
#> 12:        esg-dn3.nsc.liu.se     UP     9999
#> 13:            esg.camscma.cn     UP     9999
#> 14:            esg.lasg.ac.cn     UP     9999
#> 15:          esg1.umr-cnrm.fr     UP     9999
#> 16:          esg2.umr-cnrm.fr     UP     9999
#> 17:     esgdata.gfdl.noaa.gov     UP     9999
#> 18:    esgf-cnr.hpc.cineca.it     UP     9999
#> 19:        esgf-data.ucar.edu     UP     9999
#> 20:    esgf-data01.diasjp.net     UP     9999
#> 21:    esgf-data02.diasjp.net     UP     9999
#> 22:    esgf-data03.diasjp.net     UP     9999
#> 23:    esgf-data04.diasjp.net     UP     9999
#> 24:     esgf-data1.diasjp.net     UP     9999
#> 25:       esgf-data1.llnl.gov     UP     9999
#> 26:       esgf-data2.llnl.gov     UP     9999
#> 27:   esgf-ictp.hpc.cineca.it     UP     9999
#> 28: esgf-nimscmip6.apcc21.org     UP     9999
#> 29:         esgf-node.cmcc.it     UP     9999
#> 30:        esgf-node.ornl.gov     UP     9999
#> 31:        esgf-node2.cmcc.it     UP     9999
#> 32:           esgf.apcc21.org     UP     9999
#> 33:               esgf.bsc.es     UP     9999
#> 34:           esgf.ceda.ac.uk     UP     9999
#> 35:               esgf.dwd.de     UP     9999
#> 36:             esgf.ichec.ie     UP     9999
#> 37:        esgf.nccs.nasa.gov     UP     9999
#> 38:           esgf.nci.org.au     UP     9999
#> 39:   esgf.rcec.sinica.edu.tw     UP     9999
#> 40:             esgf1.dkrz.de     UP     9999
#> 41:             esgf2.dkrz.de     UP     9999
#> 42:             esgf3.dkrz.de     UP     9999
#> 43:     noresg.nird.sigma2.no     UP     9999
#> 44:        polaris.pknu.ac.kr     UP     9999
#> 45:         vesg.ipsl.upmc.fr     UP     9999
#> 46:           cmip.bcc.cma.cn   DOWN     9999
#> 47:      dist.nmlab.snu.ac.kr   DOWN     9999
#> 48:           dmiesgf0.dmi.dk   DOWN     9999
#> 49:       eridanus.eoc.dlr.de   DOWN     9999
#> 50:        esg-dn2.nsc.liu.se   DOWN     9999
#> 51:        esg.pik-potsdam.de   DOWN     9999
#> 52:          esgf-data.csc.fi   DOWN     9999
#> 53:              esgf.anl.gov   DOWN     9999
#>                     data_node status speed_ms
# }