summary CMIP5 files information

get_ensemble: Extract CMIP5 ensemle name from file names

extract_model: Extract CMIP5 model name from file names

CMIP5Files_info: Extract CMIP5 information from file names

CMIP5Files_filter: Filter corresponding duration or period files. Dates has been adjust from 12-01 to 01-01 for start_adj. Duplicated dates are also removed.

CMIP5Files_summary: Get the start and end information of every model

get_ensemble(files)

extract_model(files)

CMIP5Files_info(files)

CMIP5Files_summary(d)

CMIP5Files_filter(
  files,
  duration = 200,
  period = NULL,
  check_dupli = TRUE,
  verbose = 1
)

Arguments

files

CMIP5 nc files, full name path.

d

An object returned by CMIP5Files_filter or CMIP5Files_info or a data.frame at least with the columns of 'model', 'year_start_adj', 'year_end', 'file'.

duration

length of year.

period

starting and ending year, e.g. c(1850, 2005). If provided, duration will be overwritten. If duration and period are null at the same time, no filter applied.

check_dupli

Boolean. If true, duplicated date will be removed by check_dfile().

verbose

echo nc file missing info, if verbose >= 1

echo duplicated info, if verbose >= 2

Value

d_files A data.table with colnames of 'Id', 'model', 'kind', 'start', 'end', 'year_start', 'year_start_adj', 'year_end', 'file'

Note

Only one scenario per time.

See also

Examples

data(files_short)
ensemble <- get_ensemble(files_short)
model <- extract_model(files_short)
CMIP5Files_info(files_short)
#>         Id  scenario     model ensemble freq      start        end  start_adj
#>    1:    1 piControl ACCESS1-0   r1i1p1  day  300-01-01  324-12-31  300-01-01
#>    2:    2 piControl ACCESS1-0   r1i1p1  day  325-01-01  349-12-31  325-01-01
#>    3:    3 piControl ACCESS1-0   r1i1p1  day  350-01-01  374-12-31  350-01-01
#>    4:    4 piControl ACCESS1-0   r1i1p1  day  375-01-01  399-12-31  375-01-01
#>    5:    5 piControl ACCESS1-0   r1i1p1  day  400-01-01  424-12-31  400-01-01
#>   ---                                                                        
#> 2448: 2448 piControl    inmcm4   r1i1p1  day 2300-01-01 2309-12-31 2300-01-01
#> 2449: 2449 piControl    inmcm4   r1i1p1  day 2310-01-01 2319-12-31 2310-01-01
#> 2450: 2450 piControl    inmcm4   r1i1p1  day 2320-01-01 2329-12-31 2320-01-01
#> 2451: 2451 piControl    inmcm4   r1i1p1  day 2330-01-01 2339-12-31 2330-01-01
#> 2452: 2452 piControl    inmcm4   r1i1p1  day 2340-01-01 2349-12-31 2340-01-01
#>          end_adj year_start year_end len
#>    1:  324-12-31        300      324  25
#>    2:  349-12-31        325      349  25
#>    3:  374-12-31        350      374  25
#>    4:  399-12-31        375      399  25
#>    5:  424-12-31        400      424  25
#>   ---                                   
#> 2448: 2309-12-31       2300     2309  10
#> 2449: 2319-12-31       2310     2319  10
#> 2450: 2329-12-31       2320     2329  10
#> 2451: 2339-12-31       2330     2339  10
#> 2452: 2349-12-31       2340     2349  10
#>                                                             file
#>    1: tasmax_day_ACCESS1-0_piControl_r1i1p1_03000101-03241231.nc
#>    2: tasmax_day_ACCESS1-0_piControl_r1i1p1_03250101-03491231.nc
#>    3: tasmax_day_ACCESS1-0_piControl_r1i1p1_03500101-03741231.nc
#>    4: tasmax_day_ACCESS1-0_piControl_r1i1p1_03750101-03991231.nc
#>    5: tasmax_day_ACCESS1-0_piControl_r1i1p1_04000101-04241231.nc
#>   ---                                                           
#> 2448:    tasmax_day_inmcm4_piControl_r1i1p1_23000101-23091231.nc
#> 2449:    tasmax_day_inmcm4_piControl_r1i1p1_23100101-23191231.nc
#> 2450:    tasmax_day_inmcm4_piControl_r1i1p1_23200101-23291231.nc
#> 2451:    tasmax_day_inmcm4_piControl_r1i1p1_23300101-23391231.nc
#> 2452:    tasmax_day_inmcm4_piControl_r1i1p1_23400101-23491231.nc

# Get the summary infomation of CMIP5Files_filter or CMIP5Files_info
CMIP5Files_filter(files_short) %>% CMIP5Files_summary()
#> Error in loadNamespace(x): there is no package called ‘plyr’

# filter CMIP5 files
CMIP5Files_filter(files_short, duration = 200)
#> Error in loadNamespace(x): there is no package called ‘plyr’